feat: Add defineProperties module for cache store
Added a new module called defineProperties.cjs to define the properties of a cache store instance. The module exports a function called defineProperties which takes an instance of a cache store and defines the properties using Object.defineProperty. The properties include hasItem, getItem, addItem, and remItem.

The module is added to the src/mods/stores/cache/definePropertries.cjs file.

This commit addresses the need to define properties for a cache store instance and improves the overall functionality of the cache store.