feat: Add getJson method to JSON cache store
Add a new method called getJson to the JSON cache store module. This method allows fetching JSON data from the cache based on the provided URL, time-to-live (ttl), and dynamic (dinanic) properties. The method utilizes the cacheJson function from the jsonCacheComponents module to fetch the JSON data. If the data is found in the cache, it will be returned; otherwise, undefined will be returned.