feat(cacheJson): add function to prepare save schema
This commit adds a new function `prepareSave` to the `cacheJson` module in order to prepare the save schema based on the data received. The function takes into account the settings, request, and size of the data to determine the appropriate values for the TTL and URL in the schema. This allows for more flexibility in handling dynamic requests and caching large JSON responses.